Simplify /rate page and remove gamification
