--- src/__tests__/useStreak.test.ts
@@ -1,20 +1,0 @@
-import { renderHook, act } from '@testing-library/react-hooks';
-import { useStreak } from '../hooks/useStreak';
-
-describe('useStreak', () => {
-  it('increments streak once per day', () => {
-    const { result } = renderHook(() => useStreak());
-    act(() => {
-      const [, inc] = result.current;
-      inc();
-    });
-    const [state] = result.current;
-    expect(state.streak).toBe(1);
-    act(() => {
-      const [, inc] = result.current;
-      inc();
-    });
-    const [afterSecond] = result.current;
-    expect(afterSecond.streak).toBe(1);
-  });
-});
