--- src/constants/gamification.ts
@@ -1,10 +1,0 @@
-export const BADGES = [
-  { threshold: 50, name: 'Rookie Rater' },
-  { threshold: 100, name: 'Seasoned Scouter' },
-  { threshold: 500, name: 'Hall of Fame' },
-];
-
-export const LEVEL_THRESHOLDS = [
-  { level: 1, xp: 0 },
-  { level: 5, xp: 500 },
-];
